Cria robo de video e passa rendenizar video via command line do after effects
