+++ .gitignore
@@ -109,3 +109,6 @@ credentials/*.json
 content.json
 
 content/*.png
+content/*.jpg
+content/*.mov
+content/*.js
