+++ index.js
@@ -2,13 +2,15 @@ const robots = {
     input: require('./robots/input.js'),
     text: require('./robots/text.js'),
     state: require('./robots/state.js'),
-    image: require('./robots/images.js')
+    image: require('./robots/images.js'),
+    video: require('./robots/video.js')
 }
 
 async function start() {
     robots.input();
     await robots.text();
-    await robots.image();
+    await robots.image(); 
+    await robots.video();
 
     const content = robots.state.load()
     console.dir(content, { depth: null })
