+++ robots/state.js
@@ -1,9 +1,16 @@
 const fs = require('fs')
 const contentFilePath = './content.json'
+const scriptFilePath = './content/after-effects-script.js'
 
 function save(content) {
     const contentString = JSON.stringify(content)
     return fs.writeFileSync(contentFilePath, contentString)
+}
+
+function saveScript(content) {
+    const contentString = JSON.stringify(content)
+    const scriptString = `var content = ${contentString}`
+    return fs.writeFileSync(scriptFilePath, scriptString)
 }
 
 function load() {
@@ -14,5 +21,6 @@ function load() {
 
 module.exports = {
     save,
+    saveScript,
     load
 }
